Fix compiling error on macOS.

+++ src/kernel/Communicator.cc
@@ -1931,7 +1931,7 @@ int Communicator::reply_message_unreliable(struct CommConnEntry *entry)
 			.msg_name		=	entry->target->addr,
 			.msg_namelen	=	entry->target->addrlen,
 			.msg_iov		=	vectors,
-			.msg_iovlen		=	(size_t)cnt,
+			.msg_iovlen		=	cnt,
 		};
 		if (sendmsg(entry->sockfd, &message, 0) < 0)
 			return -1;
